+++ pyproject.toml
@@ -67,7 +67,7 @@ archive_to_bigquery = []
 
 check_topics_by_upd_time = []
 
-check_first_posts_for_changes = ["feedparser>=6.0", "PyMySQL>=1.1.3"]
+check_first_posts_for_changes = ["PyMySQL>=1.1.3"]
 
 communicate = []
