--- src/check_first_posts_for_changes/_utils/commons.py
@@ -1,36 +1,0 @@
-from dataclasses import dataclass, field
-from datetime import datetime
-
-
-@dataclass
-class Search:
-    topic_id: int
-
-
-@dataclass
-class RSSItem:
-    item_id: str  # URL oc comment
-    topic_id: int
-    published_at: datetime
-    updated_at: datetime
-    content: str
-    id: int | None = None
-
-
-@dataclass
-class PercentGroup:
-    n: int
-    start_percent: int
-    finish_percent: int
-    frequency: int
-    first_delay: int
-    start_num: int = 0
-    finish_num: int = 0
-    searches: list[Search] = field(default_factory=list)  # searches
-
-    def __str__(self) -> str:
-        days = f' or {int(self.frequency // 1440)} day(s)' if self.frequency >= 1440 else ''
-        return (
-            f'N{self.n: <2}: {self.start_percent}%–{self.finish_percent}%. Updated every {self.frequency} minute(s){days}. '
-            f'First delay = {self.first_delay} minutes. nums {self.start_num}-{self.finish_num}. num of searches {len(self.searches)}'
-        )
