+++ src/check_first_posts_for_changes/requirements.txt
@@ -25,8 +25,6 @@ charset-normalizer==3.4.4
     # via requests
 decorator==5.2.1
     # via retry2
-feedparser==6.0.12
-    # via la-searcher-bot
 greenlet==3.2.4 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
     # via sqlalchemy
 h11==0.16.0
@@ -61,6 +59,8 @@ pydantic-core==2.41.5
     # via pydantic
 pydantic-settings==2.12.0
     # via la-searcher-bot
+pymysql==1.1.3
+    # via la-searcher-bot
 python-dateutil==2.8.2
     # via botocore
 python-dotenv==1.2.1
@@ -77,8 +77,6 @@ retry2==0.9.5
     # via la-searcher-bot
 s3transfer==0.15.0
     # via boto3
-sgmllib3k==1.0.0
-    # via feedparser
 six==1.17.0
     # via python-dateutil
 sniffio==1.3.1
