--- tests/fixtures/rss_feed.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<feed xmlns="http://www.w3.org/2005/Atom" xml:lang="ru">
-	<link rel="self" type="application/atom+xml" href="https://lizaalert.org/forum/app.php/feed" />
-
-	<title>ЛизаАлерт : Добровольческий поисково-спасательный отряд</title>
-	<subtitle>Форум</subtitle>
-	<link href="https://lizaalert.org/forum/index.php" />
-	<updated>2026-02-25T20:53:21+03:00</updated>
-
-	<author>
-		<name><![CDATA[ЛизаАлерт : Добровольческий поисково-спасательный отряд]]></name>
-	</author>
-	<id>https://lizaalert.org/forum/app.php/feed</id>
-
-	<entry>
-		<author>
-			<name><![CDATA[Сова Александр]]></name>
-		</author>
-		<updated>2026-02-25T20:53:21+03:00</updated>
-		<published>2026-02-25T20:53:21+03:00</published>
-		<id>https://lizaalert.org/forum/viewtopic.php?p=1120245#p1120245</id>
-		<link href="https://lizaalert.org/forum/viewtopic.php?p=1120245#p1120245" />
-		<title type="html"><![CDATA[Ночной патруль • Re: г. Екатеринбург и Свердловская обл. 25.02.2026 год]]></title>
-
-		<category term="Ночной патруль" scheme="https://lizaalert.org/forum/viewforum.php?f=517"
-			label="Ночной патруль" />
-
-		<content type="html" xml:base="https://lizaalert.org/forum/viewtopic.php?p=1120245#p1120245"><![CDATA[
-Экипаж 607 Екатеринбург к ночному патрулированию приступил пилот Филин +79000000000, штурман Анна +79000000001<p>Статистика: Добавлено  <a href="https://lizaalert.org/forum/memberlist.php?mode=viewprofile&amp;u=159846">Сова Александр</a> — 25 фев 2026, 20:53</p><hr />
-]]></content>
-
-	</entry>
-
-	<entry>
-		<author>
-			<name><![CDATA[Горизонт]]></name>
-		</author>
-		<updated>2026-02-25T20:52:34+03:00</updated>
-		<published>2026-02-25T20:52:34+03:00</published>
-		<id>https://lizaalert.org/forum/viewtopic.php?p=1120244#p1120244</id>
-		<link href="https://lizaalert.org/forum/viewtopic.php?p=1120244#p1120244" />
-		<title type="html"><![CDATA[Активные поиски • Re: Пропала девочка, 9 лет, г. Смоленск]]></title>
-
-		<category term="Активные поиски" scheme="https://lizaalert.org/forum/viewforum.php?f=276"
-			label="Активные поиски" />
-
-		<content type="html" xml:base="https://lizaalert.org/forum/viewtopic.php?p=1120244#p1120244"><![CDATA[
-Горизонт в экипаже с Фенка и Почтальон. РВП 02:00 +79000000000<p>Статистика: Добавлено  <a href="https://lizaalert.org/forum/memberlist.php?mode=viewprofile&amp;u=98191">Горизонт</a> — 25 фев 2026, 20:52</p><hr />
-]]></content>
-
-	</entry>
-
-	<entry>
-		<author>
-			<name><![CDATA[ФЕНКА]]></name>
-		</author>
-		<updated>2026-02-25T20:48:05+03:00</updated>
-		<published>2026-02-25T20:48:05+03:00</published>
-		<id>https://lizaalert.org/forum/viewtopic.php?p=1120243#p1120243</id>
-		<link href="https://lizaalert.org/forum/viewtopic.php?p=1120243#p1120243" />
-		<title type="html"><![CDATA[Активные поиски • Re: Пропала девочка, 9 лет, г. Смоленск]]></title>
-
-		<category term="Активные поиски" scheme="https://lizaalert.org/forum/viewforum.php?f=276"
-			label="Активные поиски" />
-
-		<content type="html" xml:base="https://lizaalert.org/forum/viewtopic.php?p=1120243#p1120243"><![CDATA[
-Фенка а экипаже с Почтальоном и Горизонтом на авто, РВП позже +79000000000<p>Статистика: Добавлено  <a href="https://lizaalert.org/forum/memberlist.php?mode=viewprofile&amp;u=101360">ФЕНКА</a> — 25 фев 2026, 20:48</p><hr />
-]]></content>
-
-	</entry>
-
-</feed>
